+++ app/static/js/bulletin.js
