--- templates/list.html
@@ -1,16 +1,0 @@
-{% extends 'layout.html' %}
-
-{% block page %}
-<div class="container">
-  <div class="row">
-    <div class="col-xs-12">
-      {% include 'nav.html' %}
-    </div>
-  </div>
-  <div class="row">
-    <div class="col-xs-12">
-      {% include 'results.html' %}
-    </div>
-  </div>
-</div>
-{% endblock %}
